requirements.txt
pandas==2.3.1
pandas-ta==0.3.14b0
numpy
ccxt==4.3.4
yfinance==0.2.36
python-dotenv==1.0.1
matplotlib==3.8.2
